% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_d.R
\name{fit_rs_meta_d_MLE}
\alias{fit_rs_meta_d_MLE}
\title{Calculate response-specific meta-d' using Maximum likelihood estimation.}
\usage{
fit_rs_meta_d_MLE(nR_S1, nR_S2, s = 1, fncdf = pnorm, fninv = qnorm)
}
\arguments{
\item{nR_S1}{A vector. Contains the total number of responses in the first response category, conditional on presentation of S1 and S2.

If nR_S1 = c(100,50,20,10,5,1), then when stimulus S1 was presented, the subject had the following response counts:

responded S1, rating=3 : 100 times

responded S1, rating=2 : 50 times

responded S1, rating=1 : 20 times

responded S2, rating=1 : 10 times

responded S2, rating=2 : 5 times

responded S2, rating=3 : 1 time}

\item{nR_S2}{A vector. Contains the total number of responses in the second response category, conditional on presentation of S1 and S2.

The ordering of response / rating counts for S2 should be the same as it is for S1.

If nR_S2 = c(3,7,8,12,27,89), then when stimulus S2 was presented, the subject had the following response counts:

responded S1, rating=3 : 3 times

responded S1, rating=2 : 7 times

responded S1, rating=1 : 8 times

responded S2, rating=1 : 12 times

responded S2, rating=2 : 27 times

responded S2, rating=3 : 89 times}

\item{s}{A number. The ratio of standard deviations for type 1 distributions, i.e. s = sd(S1) / sd(S2). If not specified, s is set to a default value of 1.}

\item{fncdf}{A function for the CDF of the type 1 distribution. If not specified, fncdf defaults to \code{pnorm} (i.e. CDF for normal distribution).}

\item{fninv}{A function for the inverse CDF of the type 1 distribution. If not specified, fninv defaults to \code{qnorm} (i.e. inverse CDF for normal distribution).}
}
\value{
A list. In the following, let S1 and S2 represent the distributions of evidence
generated by stimulus classes S1 and S2. Then the fields is a list are as follows:
da        = mean(S2) - mean(S1), in room-mean-square(sd(S1),sd(S2)) units

t1ca      = type 1 criterion for overall data, RMS units

s         = sd(S1) / sd(S2)

meta_da_rS1   = meta-d' for S1 responses, RMS units

t1ca_rS1      = type 1 criteron for meta-d' fit for S1 responses, RMS units

t2ca_rS1      = type 2 criteria for meta-d' fit for S1 responses, RMS units

M_ratio_rS1   = meta_da_rS1 / da

M_diff_rS1    = meta_da_rS1 - da

logL_rS1      = log likelihood of meta-d' fit for S1 responses

obs_HR2_rS1   = actual type 2 hit rates for S1 responses

est_HR2_rS1   = estimated type 2 hit rates for S1 responses

obs_FAR2_rS1  = actual type 2 false alarm rates for S1 responses

est_FAR2_rS1  = estimated type 2 false alarm rates for S1 responses


meta_da_rS2   = meta-d' for S2 responses, RMS units

t1ca_rS2      = type 1 criteron for meta-d' fit for S2 responses, RMS units

t2ca_rS2      = type 2 criteria for meta-d' fit for S2 responses, RMS units

M_ratio_rS2   = meta_da_rS2 / da

M_diff_rS2    = meta_da_rS2 - da

logL_rS2      = log likelihood of meta-d' fit for S2 responses

obs_HR2_rS2   = actual type 2 hit rates for S2 responses

est_HR2_rS2   = estimated type 2 hit rates for S2 responses

obs_FAR2_rS2  = actual type 2 false alarm rates for S2 responses

est_FAR2_rS2  = estimated type 2 false alarm rates for S2 responses

S1units contains same parameters in sd(S1) units.
}
\description{
Given data from an experiment where an observer discriminates between two stimulus
alternatives on every trial and provides confidence ratings, provides a RESPONSE-SPECIFIC type 2
SDT analysis of the data. i.e. meta-d' is computed separately for each response type.
}
