% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_d.R
\name{trials2counts}
\alias{trials2counts}
\title{Converts trial by trial experimental information into response counts.}
\usage{
trials2counts(stimID, response, rating, nRatings, padCells = FALSE,
  padAmount = 1/(2 * nRatings))
}
\arguments{
\item{stimID}{1xN vector. stimID[i] = 0 --> stimulus on i'th trial was S1; stimID[i] = 1 --> stimulus on i'th trial was S2.}

\item{response}{1xN vector. response[i] = 0 --> response on i'th trial was "S1"; response[i] = 1 --> response on i'th trial was "S2".}

\item{rating}{1xN vector. rating(i) = X --> rating on i'th trial was X; X must be in the range 1 <= X <= nRatings.}

\item{nRatings}{A number. Total number of available subjective ratings available for the subject. e.g. if subject can rate confidence on a scale of 1-4, then nRatings = 4.}

\item{padCells}{If set to TRUE, each response count in the output has the value of padAmount added to it. Padding cells is desirable if trial counts of 0 interfere with model fitting. If set to FALSE, trial counts are not manipulated and 0s may be present in the response count output.}

\item{padAmount}{The value to add to each response count if padCells == TRUE. Default value is 1/(2*nRatings).}
}
\value{
A list contains two vectors, nR_S1 and nR_S2. These are vectors containing the total number of responses in each response category, conditional on presentation of S1 and S2.

    If nR_S1 = c(100,50,20,10,5,1), then when stimulus S1 was presented, the subject had the following response counts:

    responded S1, rating=3 : 100 times

    responded S1, rating=2 : 50 times

    responded S1, rating=1 : 20 times

    responded S2, rating=1 : 10 times

    responded S2, rating=2 : 5 times

    responded S2, rating=3 : 1 time
}
\description{
Given data from an experiment where an observer discriminates between two stimulus alternatives on every trial and provides confidence ratings, converts trial by trial experimental information for N trials into response counts.
}
